Profile page

+++ public/javascripts/user-profile.js
@@ -1,24 +1,26 @@
 const followButton = document.querySelector('.green-btn-follow')
-followButton.addEventListener('click', (e) => {
-    e.preventDefault();
-    if (followButton.innerHTML === "Follow") {
-        fetch(`/users/${followButton.dataset.value}`, {
-            method: "PATCH",
-        }).then(function(followButton) {
-            document.querySelector(".green-btn-follow").innerHTML = "Following"
-        }).catch((error) => {
-            console.error(error)
-        })
-    } else {
-        fetch(`/users/${followButton.dataset.value}`, {
-            method: "DELETE",
-        }).then(function(followButton) {
-            document.querySelector(".green-btn-follow").innerHTML = "Follow"
-        }).catch((error) => {
-            console.error(error)
-        })
-    }
-})
+if(followButton){
+    followButton.addEventListener('click', (e) => {
+        e.preventDefault();
+        if (followButton.innerHTML === "Follow") {
+            fetch(`/users/${followButton.dataset.value}`, {
+                method: "PATCH",
+            }).then(function(followButton) {
+                document.querySelector(".green-btn-follow").innerHTML = "Following"
+            }).catch((error) => {
+                console.error(error)
+            })
+        } else {
+            fetch(`/users/${followButton.dataset.value}`, {
+                method: "DELETE",
+            }).then(function(followButton) {
+                document.querySelector(".green-btn-follow").innerHTML = "Follow"
+            }).catch((error) => {
+                console.error(error)
+            })
+        }
+    })
+}
